Menus
Added Menus to the Engine. This includes a Menu class that holds all
informationa and items in a menu, and a MenuManager class that manages
the Menus.
Also on a unrelated note this update fixes #3. Added upper, lower, left
and right bound properties to the Sprite to easily get where its extreme
points are and compare againt limitboxes. They are from centre of Sprite
as Sprites position is now from its centre. Still havent solved Bounces
( #4)though =(